Move the engine functions into the diplomacy.engine namespace
